Extract `collection/collection_ops.rs`
